--- src/Renci.SshNet.Tests.NET35/Classes/ForwardedPortLocal.NET35.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Renci.SshNet.Tests.Common;
-
-namespace Renci.SshNet.Tests.SshClientTests
-{
-    public partial class ForwardedPortLocalTest : TestBase
-    {
-    }
-}
