--- src/Renci.SshNet.Tests/Classes/ShellStream.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Renci.SshNet.Tests.Common;
-
-namespace Renci.SshNet.Tests.Classes
-{
-    /// <summary>
-    /// Contains operation for working with SSH Shell.
-    /// </summary>
-    [TestClass]
-    public class ShellStreamTest : TestBase
-    {
-    }
-}
