--- src/Renci.SshNet.Tests/Properties/AssemblyInfo.cs
@@ -1,5 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("SSH.NET Tests .NET 4.0")]
-[assembly: Guid("defb66b2-a4b9-40ca-b205-2183d84b7323")]
