--- src/Renci.SshNet.Tests/authorized_keys
@@ -1,4 +1,0 @@
-ssh-rsa AAAAB3NzaC1yc2EAAAABIwAAAQEA8ZsD8jgH7ySXnd4dprEFFkJ+zs1ne81E8febjR8hekiKrc+D9GjeEykchy0PsrsBrf2F8J3GtLNISVerkl+EnVh1E0pi9Vllc1vcpQHvBeNPG/jl9Dnum/DcaaMKLot+ARXjBuMX/xJVfnlgkvfpBvoq4QzQ6E27rOMbcPlvrN7KeYbaorsTpDMUraHX5u99P1evvw7sOuNl3lc9YpIgmjHT6dWUGTqBx9T2SEKXBRVk2NNbd2UctQkr5BJn1gGehA+1KYS30FMiCSN8F8ZOlpYj1+K6xzs0srq1z30LailIyy7QmYp739BJlOF3cVywdcFXGre35H2cJylcNV2e+Q== rsa_key_without_pass
-ssh-rsa AAAAB3NzaC1yc2EAAAABIwAAAQEAuD5SKSM3RAO/h8z9x1RB9bhLbBkophrYVHYG8MdNwQc/iXb4IZ7PlKR2ZQovIYzHdMI5zGEuOsxJnLjxoV7NSUHlHEoGbRYX72rR3S3ZORUArLTJ9TwGZhR4TZRztJto7l4M5LGDUy4YMNN9kpcyT/aqv8v3ZcTB0UthZDXUHrrmrlktPAC+JagUjhESbtq+FAuiSUWd9/SXnAmLya7LkWANw2ScMpeSyXmEvLD+X0eapNtuwXltjRW4PoKS5fWMhiXnT61NXmgXDccrUHjWrOa1oeNjmhOuIGaDpoRtkBH8lCFQdRePgvcBIRnU9D7sjuBjZUAPqin+fDll2ISy2w== rsa_key_with_pass (tester)
-ssh-dss AAAAB3NzaC1kc3MAAACBAN1CuvAa+SfAP6lv0am/+8PGPzKoVSDRxS4L+2SObe76VCPzpHB8DxEk5X44/ga2+z3M5KzGK6CKegH+4SJP7qNse/K8l4mnZ+wNYvH+1VFfGmLw3B5t9vjnlHTBVVl2qcUqwBkutABPnz0RS0uDJUAVBtQ/kXPqnHwfwAbSqXJtAAAAFQC5+1DsZ2xUe3w32ZKtj3O2QlXoKQAAAIEA1Ep/297O7PVoZSdQZYgsc7T1KJscrbnl+Su7j//O8Jt1rBZJk6hT68yemUtSE1pUyhLJgT9Ng5m+9sb1S0++vzBkZ00+Km8hcgHGWmv4UfSadLHSqVpHLc8SgNECbDqaxELNVhXe9nxlMRR236mG6u76mfRb0S6DMIweKRVWEkwAAACASHB5nNlWhpqvungn+otYyGKoVoATkRdRfSiUo3fb0mJLTmxrEhEH4UOFA/UFdQ0TYXXW0wRX/amynT4iTEXAx0FjxUNK5ryxloeoXiXEH4FK7D8RtJO11UsaRrN+nqWiSTVAehl6tzoMCPargGcWeFlAZZnPGN76OD9J1GiKZj8= dsa_key_without_pass
-ssh-dss AAAAB3NzaC1kc3MAAACBAJU0pDZQHna2HaaGHsL1D0pZlnh/tM7EYAcGQcHqNRUVPLXbcy94ZdF0R9V1qCA9u7at/EEqdtfseXLbjbxLxccJM6sjCFn7GKNvc6N7w7PpAeUbZBdHdszOkNolyqGs+/45Oj0tGHLTLk2rRyOTeBagpJS1PlZQCnVbT8o/gjtvAAAAFQDUe90v8pdAZhtO84F19cRUMrWoqQAAAIBDZwBGOxMRy7nKc0f6aWNh0JX5NnTCuEiXmD6BmBV/2k9J0VJYgiTZw4mzmF4VzrV7aI+XCQz52N8od7cp76O1DXweO0cIqLkT2xLGuoBtUSpHiSqRnHUsL9BDI8s3bQ5fSCb0IwF6d90tArDX/d9mzYqx+Vvp+4ajSiSiyzaGUQAAAIB3vFz1T9jGAZ3hBPkFQ9zl0gniF2EQ/w1UlnHb7ABUyBigBxtnTN0+7yfNwFVfQS/rKxhT6muzYgQB3e1masvv0ky/vstnh5n7EZmLDLEWwIfRefbPmyfLCKK+23QmMZgPg/SGuLmwagw711ReiLyBqn5YiP8quXcDdnReJCJPig== dsa_key_with_pass (tester)
